refactor(cli): split large command modules
